That's my take on this issues, check my TODOs
